# docs/openapi/spec.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "All the information needed to talk to the API.\n\nTo raise see anything wrong? Raise an [issue](https://github.com/mailchain/mailchain/issues)",
    "title": "Mailchain API",
    "version": "0.0.1"
  },
  "basePath": "/api",
  "paths": {
    "/addresses": {
      "get": {
        "description": "Get all address that this user has access to. The addresses can be used to send or receive messages.",
        "tags": [
          "Addresses"
        ],
        "summary": "Get addresses.",
        "operationId": "GetAddresses",
        "responses": {
          "200": {
            "$ref": "#/responses/GetAddressesResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          }
        }
      }
    },
...
    "/nameservice/address/{address}/resolve?network={network}\u0026protocol={protocol}": {
      "get": {
        "description": "Get name for supplied address. The name is typically a human-readable value that can be used in place of the address.\nResolve will query the protocol's name service to find the human-readable name for the supplied address.",
        "tags": [
          "ResolveName",
          "NameService"
        ],
        "summary": "Resolve Address Against Name Service",
        "operationId": "GetResolveAddress",
        "parameters": [
          {
            "type": "string",
            "example": "0x4ad2b251246aafc2f3bdf3b690de3bf906622c51",
            "x-go-name": "Address",
            "description": "name to query to get address for",
            "name": "address",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "mainnet",
              "goerli",
              "ropsten",
              "rinkeby",
              "local"
            ],
            "type": "string",
            "example": "goerli",
            "x-go-name": "Network",
            "description": "Network for the name to resolve",
            "name": "network",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "ethereum"
            ],
            "type": "string",
            "example": "ethereum",
            "x-go-name": "Protocol",
            "description": "Protocol for the name to resolve",
            "name": "protocol",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetResolveAddressResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          }
        }
      }
    },
    "/nameservice/name/{domain-name}/resolve?network={network}\u0026protocol={protocol}": {
      "get": {
        "description": "Get address for supplied name. The name is typically a human-readable value that can be used in place of the address.\nResolve will query the protocol's name service to find the address for supplied human-readable name.",
        "tags": [
          "ResolveName",
          "NameService"
        ],
        "summary": "Resolve Name Against Name Service",
        "operationId": "GetResolveName",
        "parameters": [
          {
            "type": "string",
            "example": "mailchain.eth",
            "x-go-name": "Name",
            "description": "name to query to get address for",
            "name": "domain-name",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "mainnet",
              "goerli",
              "ropsten",
              "rinkeby",
              "local"
            ],
            "type": "string",
            "example": "goerli",
            "x-go-name": "Network",
            "description": "Network for the name to resolve",
            "name": "network",
            "in": "path",
            "required": true
          },
          {
            "enum": [
              "ethereum"
            ],
            "type": "string",
            "example": "ethereum",
            "x-go-name": "Protocol",
            "description": "Protocol for the name to resolve",
            "name": "protocol",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetResolveNameResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          }
        }
      }
    },
    "/protocols": {
      "get": {
        "description": "Get all networks for each protocol that is enabled.",
        "tags": [
          "protocols"
        ],
        "summary": "Get protocols and the networks.",
        "operationId": "GetProtocols",
        "responses": {
          "200": {
            "$ref": "#/responses/GetProtocolsResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          }
        }
      }
    },
    "/public-key": {
      "get": {
        "description": "Get the public key.",
        "tags": [
          "PublicKey"
        ],
        "summary": "Get public key from an address.",
        "operationId": "GetPublicKey",
        "parameters": [
          {
            "pattern": "0x[a-fA-F0-9]{40}",
            "type": "string",
            "example": "0xde0b295669a9fd93d5f28d9ec85e40f4cb697bae",
            "x-go-name": "Address",
            "description": "Address to to use when performing public key lookup.",
            "name": "address",
            "in": "query",
            "required": true
          },
          {
            "enum": [
              "mainnet",
              "goerli",
              "ropsten",
              "rinkeby",
              "local"
            ],
            "type": "string",
            "example": "goerli",
            "x-go-name": "Network",
            "description": "Network to use when performing public key lookup.",
            "name": "network",
            "in": "query",
            "required": true
          },
          {
            "enum": [
              "ethereum"
            ],
            "type": "string",
            "example": "ethereum",
            "x-go-name": "Protocol",
            "description": "Protocol to use when performing public key lookup.",
            "name": "protocol",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetPublicKeyResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundError"
          },
          "422": {
            "$ref": "#/responses/ValidationError"
          }
        }
      }
    }
  },
  "definitions": {
...
    "GetProtocolsProtocol": {
      "type": "object",
      "properties": {
        "name": {
          "description": "in: body",
          "type": "string",
          "x-go-name": "Name"
        },
        "networks": {
          "description": "in: body",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Networks"
        }
      },
      "x-go-package": "github.com/mailchain/mailchain/cmd/mailchain/internal/http/handlers"
    },
    "GetPublicKeyResponseBody": {
      "description": "GetBody body response",
      "type": "object",
      "required": [
        "public_key"
      ],
      "properties": {
        "public_key": {
          "description": "The public key",
          "type": "string",
          "x-go-name": "PublicKey",
          "example": "0x79964e63752465973b6b3c610d8ac773fc7ce04f5d1ba599ba8768fb44cef525176f81d3c7603d5a2e466bc96da7b2443bef01b78059a98f45d5c440ca379463"
        }
      },
      "x-go-package": "github.com/mailchain/mailchain/cmd/mailchain/internal/http/handlers"
    },
...
    "GetResolveAddressResponseBody": {
      "description": "GetBody body response",
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "description": "The resolved name",
          "type": "string",
          "x-go-name": "Name",
          "example": "mailchain.eth"
        }
      },
      "x-go-package": "github.com/mailchain/mailchain/cmd/mailchain/internal/http/handlers"
    },
    "GetResolveNameResponseBody": {
      "description": "GetBody body response",
      "type": "object",
      "required": [
        "address"
      ],
      "properties": {
        "address": {
          "description": "The resolved address",
          "type": "string",
          "x-go-name": "Address",
          "example": "0x4ad2b251246aafc2f3bdf3b690de3bf906622c51"
        }
      },
      "x-go-package": "github.com/mailchain/mailchain/cmd/mailchain/internal/http/handlers"
    },
...
  },
  "responses": {
    "GetAddressesResponse": {
      "description": "GetAddressesResponse Holds the response messages",
      "schema": {
        "type": "array",
        "items": {
          "type": "string"
        }
      }
    },
...
    "GetProtocolsResponse": {
      "description": "GetProtocolsResponse Holds the response messages",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/GetProtocolsProtocol"
        }
      }
    },
    "GetPublicKeyResponse": {
      "description": "GetPublicKeyResponse public key from address response",
      "schema": {
        "$ref": "#/definitions/GetPublicKeyResponseBody"
      }
    },
...
    "GetResolveAddressResponse": {
      "description": "GetResolveAddressResponse address of resolved name",
      "schema": {
        "$ref": "#/definitions/GetResolveAddressResponseBody"
      }
    },
    "GetResolveNameResponse": {
      "description": "GetResolveNameResponse address of resolved name",
      "schema": {
        "$ref": "#/definitions/GetResolveNameResponseBody"
      }
    },
    "NotFoundError": {
      "description": "Description of an error.",
      "headers": {
        "code": {
          "type": "string",
          "example": "404",
          "description": "Code describing the error"
        },
        "message": {
          "type": "string",
          "example": "Not found.",
          "description": "Description of the error"
        }
      }
    },
...
    "ValidationError": {
      "description": "Description of an error.",
      "headers": {
        "code": {
          "type": "string",
          "example": "422",
          "description": "Code describing the error"
        },
        "message": {
          "type": "string",
          "example": "Response to invalid input",
          "description": "Description of the error"
        }
      }
    }
  }
}